requirements/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-ckeditor==6.0.0
django-jalali==4.1.2
django-js-asset==1.2.2
django-nested-admin==3.3.3
gunicorn==20.1.0
jdatetime==3.6.2
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-monkey-business==1.0.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
